+++ blob/gulpfile.js
@@ -5,7 +5,7 @@ const version = require("./package.json").version;
 const zipFileName = `azurestoragejs.blob-${version}.zip`;
 
 gulp.task("zip", () => {
-  gulp
+  return gulp
     .src([
       "browser/azure-storage.blob.js",
       "browser/azure-storage.blob.min.js",